fix(default-layout): ensure mutation observers are regenerated whenever returning to recent searches
