+++ packages/@sanity/default-layout/src/navbar/search/utils/isTermsSearchable.ts
@@ -1,0 +1,5 @@
+import type {SearchTerms} from '@sanity/base'
+
+export function isTermsSearchable(terms: SearchTerms): boolean {
+  return terms.query !== '' || !!terms.types.length
+}
